Add initial support for what-if scenarios
Clicking a bar will cycle through the possible outcomes of the game. Textures are added to the bars to indicate their "what if?" status. Scores are updated (although not yet visible to the user), and the rows re-sort to reflect the new standings.

Other parts of the interface still need to be updated.