Added support for choosing what-if winners from the game finder panel.
Clicking a team name in the game finder panel will toggle a what-if scenario in which that team wins the game. This change includes rewriting the code for the game finder panel so that the DOM elements are built using d3 instead of a string (which fits in better with the rest of the design anyway).