Refactor game highlighting logic
This adds a class for managing which game is currently highlighted. The new approach eliminates bugs which appeared when trying to highlight games while the bars were moving. It also removes the limitation that prevented tooltips from reappearing when the games are sorted by confidence and someone clicks on a pick bar.