Change data stored in All Games list items
For the spans holding the team names, store an object holding both the game object and the team object. This makes it easier to avoid redundant code when creating the spans for each team, and it simplifies some of the logic in assigning what-if scenarios. This is a preliminary step to including four teams for the championship game in the All Games list.