Give player names/scores their own dedicated space
Previously, the names/scores were drawn over the bars at the left end of
the display, which had a few issues:

- It was visually cluttered.
- A slightly opaque background had to drawn behind the name/score to
  help its visibility when a gray bar was behind it.
- The name/score was drawn towards the top of the bar area to minimize
  the amount of overlap with bars. For years with many players (such as
  2021 with 16 players), the bar area was so short that the top of the
  name/score's opaque background was truncated because it overflowed the
  player's row into the row above.

The new method dedicates space to the names/scores at the left end,
squeezing the bar region horizontally to free up that space. The bars
are wide enough that making them slightly narrower isn't a big loss.