Add lines representing the point totals
These lines serve as a built-in bar chart. The goal is to make it more
obvious that, for example, first and second place are really close and
then there's a big drop to third place.

The lines appear below the row of bars. I bottom-justified the names so
that they look more attached to the lines. And then to help with the
balance of things, I also bottom-justified all the game bars. This is a
pretty significant visual departure from the old style where the game
bars were middle-justified, but I think it works better this way now
that the names are bottom-justified. It might also make it slightly
easier to distinguish the differences in bar heights, since the
difference is now fully represented on one end of the bar instead of
being split across both ends.

The downside of using a line below the bars is that it takes up a couple
pixels of vertical space, and vertical space is precious in this
visualization. I also considered an alternative solution where the width
of the game bars would be scaled to represent the player's point total.
Because this would throw off the vertical alignment of bars across
players, it couldn't be the default setting. Maybe it would happen only
when hovering over a point total, or maybe I'd add a new button to
enable it. I liked the lines better because they could be present at all
times without requiring user action to display them.